Create biosample-leasing-token.sol
Introducing BioNFTv2
BioNFT version 2 can build on the ERC-4907 token standard to include additional features for leasing and expiration of the permission to access the biological samples represented by the tokens. Here are some possible ideas for how these features could be implemented:

Leasing: BioNFT version 2 can include a leasing feature that allows the token owner to lease the permission to access the biological sample to another user for a limited time period. The lease would be represented by a new token that is minted and transferred to the lessee. During the lease period, the lessee would have access to the biological sample, but the ownership of the original BioNFT would remain with the lessor. Once the lease period expires, the permission to access the biological sample would revert back to the original owner.

Expiration: BioNFT version 2 can also include an expiration feature that allows the permission to access the biological sample to expire after a certain time period. This would enable the original owner to set a specific duration for the permission to be valid, after which it would automatically expire. The expiration date could be set at the time the BioNFT is created, or it could be updated later by the owner. Once the permission has expired, the token could still exist, but it would no longer grant access to the biological sample.

These additional features would further enhance the utility of BioNFTs for managing access to biological samples for research purposes. By allowing the leasing of permissions and setting expiration dates, the system could be more flexible and adaptable to the needs of researchers and sample owners. It would also provide additional safeguards to protect the privacy and security of the biological samples being used in research.