Update biosample-leasing-token.sol
This contract tokenizes a biosample collection tube using the ERC4907 standard, which allows the token to be leased for a given time. It includes functions based on GDPR data privacy laws, such as `giveConsent`, `revokeConsent`, `grantAccessPermission`, `revokeAccessPermission`, `grantComputePermission`, and `revokeComputePermission`. It also includes the `eraseBiodata` function to erase the biodata associated with the token. The `claimReward` function allows the owner of the token to claim a reward in the form of an ERC20 token for leasing the token. The token is identified by a unique alphanumeric serial number called "BiosampleID".